Move raw_task back to public.
